Change the way it highlights words
Previously if you type a small word that happens to appear in the
middle of a long word, it would abandon the long word and you’d have
to start again. Now whenever a character is typed it just highlights
the next character of any matching word. Typing a wrong letter or a
letter from a different word doesn’t remove the highlight.